code/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
web3==6.15.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 